Revamp project structure
