+++ SiskamlingBot.go
@@ -2,16 +2,10 @@ package main
 
 import (
 	"SiskamlingBot/bot"
-	"SiskamlingBot/bot/handler/command"
-	"SiskamlingBot/bot/handler/listener/metrics"
-	"SiskamlingBot/bot/handler/listener/picture"
-	"SiskamlingBot/bot/handler/listener/username"
+	"SiskamlingBot/bot/helper/config"
 	"SiskamlingBot/bot/helper/database"
-	"SiskamlingBot/bot/helper/telegram"
 	"github.com/PaulSonOfLars/gotgbot/v2"
 	"github.com/PaulSonOfLars/gotgbot/v2/ext"
-	"github.com/PaulSonOfLars/gotgbot/v2/ext/handlers"
-	"github.com/PaulSonOfLars/gotgbot/v2/ext/handlers/filters"
 	"log"
 	"net/http"
 	"runtime"
@@ -25,24 +19,12 @@ func init() {
 	runtime.GOMAXPROCS(runtime.NumCPU())
 }
 
-func addHandler(b *gotgbot.Bot, dispatcher *ext.Dispatcher) {
-	dispatcher.AddHandler(handlers.NewCommand("ping", command.Ping))
-
-	dispatcher.AddHandlerToGroup(handlers.NewMessage(filters.All, metrics.ChatMetrics), 0)
-	dispatcher.AddHandlerToGroup(handlers.NewMessage(filters.All, metrics.UsernameMetrics), 0)
-	dispatcher.AddHandlerToGroup(handlers.NewMessage(telegram.UsernameAndGroupFilter, username.Username), 0)
-	dispatcher.AddHandlerToGroup(handlers.NewMessage(telegram.ProfileAndGroupFilter(b), picture.Picture), 0)
-
-	dispatcher.AddHandlerToGroup(handlers.NewCallback(filters.Prefix("username("), username.UsernameCB), 1)
-	dispatcher.AddHandlerToGroup(handlers.NewCallback(filters.Prefix("picture("), picture.PictureCB), 1)
-}
-
 func receiveUpdates(b *gotgbot.Bot, updater ext.Updater) {
-	if bot.Config.WebhookURL != "" {
+	if config.Config.WebhookURL != "" {
 		webhook := ext.WebhookOpts{
-			Listen:  bot.Config.WebhookListen,
-			Port:    bot.Config.WebhookPort,
-			URLPath: bot.Config.WebhookPath + b.Token,
+			Listen:  config.Config.WebhookListen,
+			Port:    config.Config.WebhookPort,
+			URLPath: config.Config.WebhookPath + b.Token,
 		}
 
 		// Delete webhook before starting the bot.
@@ -56,7 +38,7 @@ func receiveUpdates(b *gotgbot.Bot, updater ext.Updater) {
 			panic("failed to start webhook: " + err.Error())
 		}
 
-		ok, err := b.SetWebhook(bot.Config.WebhookURL+bot.Config.WebhookPath+b.Token, &gotgbot.SetWebhookOpts{MaxConnections: 40})
+		ok, err := b.SetWebhook(config.Config.WebhookURL+config.Config.WebhookPath+b.Token, &gotgbot.SetWebhookOpts{MaxConnections: 40})
 		if err != nil {
 			panic("failed to start webhook: " + err.Error())
 		}
@@ -82,13 +64,13 @@ func receiveUpdates(b *gotgbot.Bot, updater ext.Updater) {
 
 func main() {
 	// Init config.
-	bot.NewConfig()
+	config.NewConfig()
 
 	// Connect to DB.
 	database.NewMongo()
 
 	// Create bot from environment value.
-	b, err := gotgbot.NewBot(bot.Config.BotAPIKey, &gotgbot.BotOpts{
+	b, err := gotgbot.NewBot(config.Config.BotAPIKey, &gotgbot.BotOpts{
 		Client:      http.Client{},
 		GetTimeout:  gotgbot.DefaultGetTimeout,
 		PostTimeout: gotgbot.DefaultPostTimeout,
@@ -102,7 +84,7 @@ func main() {
 	dispatcher := updater.Dispatcher
 
 	// Add handler to dispatcher.
-	addHandler(b, dispatcher)
+	bot.AddHandler(b, dispatcher)
 
 	// Start receiving updates.
 	receiveUpdates(b, updater)
