+++ bot/handler/command/start.go
@@ -1,0 +1,1 @@
+package command
